Debug drawing class
