+++ src/Video/DebugDrawing.cpp
@@ -1,0 +1,3 @@
+#include "DebugDrawing.hpp"
+
+using namespace Video;
